min opengl constant cleanup & case insensitive commands

+++ src/main/java/cc/hyperium/SplashProgress.java
@@ -69,7 +69,7 @@ public class SplashProgress {
         framebuffer.bindFramebuffer(false);
         GlStateManager.matrixMode(GL11.GL_PROJECTION);
         GlStateManager.loadIdentity();
-        GlStateManager.ortho(0.0D, (double) scaledresolution.getScaledWidth(), (double) scaledresolution.getScaledHeight(), 0.0D, 1000.0D, 3000.0D);
+        GlStateManager.ortho(0.0D, scaledresolution.getScaledWidth(), scaledresolution.getScaledHeight(), 0.0D, 1000.0D, 3000.0D);
         GlStateManager.matrixMode(GL11.GL_MODELVIEW);
         GlStateManager.loadIdentity();
         GlStateManager.translate(0.0F, 0.0F, -2000.0F);
@@ -107,7 +107,7 @@ public class SplashProgress {
 
         ScaledResolution sr = new ScaledResolution(Minecraft.getMinecraft());
 
-        double nProgress = (double) PROGRESS;
+        double nProgress = PROGRESS;
         double calc = (nProgress / MAX) * sr.getScaledWidth();
 
         Gui.drawRect(0, sr.getScaledHeight() - 35, sr.getScaledWidth(), sr.getScaledHeight(), new Color(0, 0, 0, 50).getRGB());
